py-requirements-dev.txt
appdirs==1.4.3
appnope==0.1.0
arrow==0.15.5
asgiref==3.2.3
attrs==19.3.0
backcall==0.1.0
black==19.10b0
blessed==1.17.2
blessings==1.7
bpython==0.19
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
curtsies==0.3.4
decorator==4.4.1
Django==3.0.3
django-admin-easy==0.6
django-cacheops==4.2
django-cors-headers==3.2.1
django-csv-export-view==1.1.0
django-extensions==2.2.8
django-grappelli==2.14.1
django-picklefield==2.1.1
django-q==1.1.0
django-ranged-response==0.2.0
django-redis==4.11.0
django-reversion==3.0.6
django-simple-captcha==0.5.12
django-summernote==0.8.11.6
django-timezone-field==4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
entrypoints==0.3
flake8==3.7.9
funcy==1.14
greenlet==0.4.17
httpie==2.3.0
idna==2.9
ipdb==0.12.3
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.16.0
mccabe==0.6.1
parso==0.6.1
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.3
psycopg2-binary==2.8.4
PTable==0.9.2
ptipython==1.0.1
ptpython==3.0.5
ptyprocess==0.6.0
py-cricket==1.0.9
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.5.2
PyJWT==1.7.1
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
regex==2020.7.14
requests==2.23.0
requests-toolbelt==0.9.1
six==1.14.0
sqlparse==0.3.0
toml==0.10.1
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.8
wcwidth==0.1.8
